Added directory checks for adding to path

+++ zshrc
@@ -74,4 +74,11 @@ else
 export NVM_DIR="$HOME/.nvm"
 [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
 [ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"  # This loads nvm bash_completion
-export PATH="/opt/homebrew/opt/mysql-client/bin:$PATH" ## from brew install mysql-client
+
+if [[ -e "/opt/homebrew/opt/mysql-client/bin" ]]; then
+  export PATH="/opt/homebrew/opt/mysql-client/bin:$PATH" ## from brew install mysql-client
+fi
+
+if [[ -e "$HOME/.local/bin" ]]; then
+  export PATH="$HOME/.local/bin:$PATH"
+fi
